separate into multiple files, remove unused consts
