+++ .vscode/cspell.json
@@ -1,7 +1,5 @@
 {
-    "ignorePaths": [
-        "src/views/GoTV/twitchLanguageCodes.ts"
-    ],
+    "ignorePaths": ["src/views/GoTV/twitchLanguageCodes.ts"],
     "words": [
         "abcdefghjklmnopqrstuvwxyz",
         "AFAICT",
@@ -182,7 +180,8 @@
         "yike",
         "yomi",
         "zenmode",
-        "zoomable"
+        "zoomable",
+        "testid"
     ],
     "language": "en,en-GB"
 }
