+++ src/components/ChallengeModal/ChallengeModal.tsx
@@ -24,19 +24,13 @@ import { OgsResizeDetector } from "@/components/OgsResizeDetector";
 import { browserHistory } from "@/lib/ogsHistory";
 import { _, pgettext, interpolate, llm_pgettext } from "@/lib/translate";
 import { post, del } from "@/lib/requests";
-import { Modal, openModal } from "@/components/Modal";
+import { Modal } from "@/components/Modal";
 import { socket } from "@/lib/sockets";
-import { rankString, amateurRanks, allRanks, rankSelectorIndexToText } from "@/lib/rank_utils";
-import { CreatedChallengeInfo, RuleSet } from "@/lib/types";
-import { errorLogger, errorAlerter, rulesText, dup } from "@/lib/misc";
+import { rankString, amateurRanks, allRanks } from "@/lib/rank_utils";
+import { CreatedChallengeInfo } from "@/lib/types";
+import { errorLogger, errorAlerter, dup } from "@/lib/misc";
 import { PlayerIcon } from "@/components/PlayerIcon";
-import {
-    shortShortTimeControl,
-    isLiveGame,
-    TimeControlPicker,
-    timeControlSystemText,
-    TimeControl,
-} from "@/components/TimeControl";
+import { isLiveGame, TimeControlPicker, TimeControl } from "@/components/TimeControl";
 import { sfx } from "@/lib/sfx";
 import {
     notification_manager,
@@ -44,15 +38,6 @@ import {
 } from "@/components/Notifications/NotificationManager";
 import { one_bot, bot_count, bots_list, getAcceptableTimeSetting, Bot } from "@/lib/bots";
 import { goban_view_mode } from "@/views/Game/util";
-import {
-    GobanRenderer,
-    GobanEngineConfig,
-    GobanEngineInitialState,
-    GobanEnginePlayerEntry,
-    JGOFTimeControl,
-    JGOFTimeControlSpeed,
-    JGOFTimeControlSystem,
-} from "goban";
 
 import { copyChallengeLinkURL } from "@/components/ChallengeLinkButton";
 
@@ -64,76 +49,18 @@ import {
 } from "@/components/TimeControl/TimeControlUpdates";
 import { SPEED_OPTIONS } from "@/views/Play/SPEED_OPTIONS";
 import Select from "react-select";
-
-export type ChallengeDetails = rest_api.ChallengeDetails;
-
-type ChallengeModes = "open" | "computer" | "player" | "demo";
-
-interface Events {}
-
-export interface ChallengeModalProperties {
-    mode: ChallengeModes;
-    game_record_mode?: boolean /* when true, if mode === "demo", we will create a game instance instead of a review instance */;
-    playerId?: number;
-    initialState?: any;
-    config?: ChallengeModalConfig;
-    autoCreate?: boolean;
-    playersList?: Array<{ name: string; rank: number }>;
-    tournamentRecordId?: number;
-    tournamentRecordRoundId?: number;
-    libraryCollectionId?: number;
-    created?: (c: CreatedChallengeInfo) => void;
-}
-
-/* These rejection details come from gtp2ogs and allows bots to
- * be clear about why a challenge is being rejected. */
-export interface RejectionDetails {
-    rejection_code:
-        | "blacklisted"
-        | "board_size_not_square"
-        | "board_size_not_allowed"
-        | "handicap_not_allowed"
-        | "unranked_not_allowed"
-        | "ranked_not_allowed"
-        | "blitz_not_allowed"
-        | "too_many_blitz_games"
-        | "live_not_allowed"
-        | "too_many_live_games"
-        | "correspondence_not_allowed"
-        | "too_many_correspondence_games"
-        | "time_control_system_not_allowed"
-        | "time_increment_out_of_range"
-        | "period_time_out_of_range"
-        | "periods_out_of_range"
-        | "main_time_out_of_range"
-        | "max_time_out_of_range"
-        | "per_move_time_out_of_range"
-        | "player_rank_out_of_range"
-        | "not_accepting_new_challenges"
-        | "too_many_games_for_player"
-        | "komi_out_of_range";
-    details: {
-        [key: string]: any;
-    };
-}
-
-interface PreferredSettingOption {
-    value: number; // index of the setting
-    label: string; // challenge text description
-    setting: ChallengeDetails;
-}
+import {
+    rejectionDetailsToMessage,
+    challenge_text_description,
+} from "@/components/ChallengeModal/ChallengeModal.utils";
+import {
+    ChallengeDetails,
+    ChallengeModalProperties,
+    PreferredSettingOption,
+    RejectionDetails,
+} from "@/components/ChallengeModal/ChallengeModal.types";
 
 /* Constants  */
-
-const negKomiRanges: number[] = [];
-const posKomiRanges: number[] = [];
-const maxKomi = 36.5;
-for (let komi = 0.0; komi <= maxKomi; komi += 0.5) {
-    if (komi - maxKomi !== 0.0) {
-        negKomiRanges.push(komi - maxKomi);
-    }
-    posKomiRanges.push(komi);
-}
 
 const handicapRanges: number[] = [];
 for (let i = 1; i <= 36; ++i) {
@@ -154,7 +81,7 @@ const standard_board_sizes: { [k: string]: string | undefined } = {
     "5x13": "5x13",
 };
 
-export class ChallengeModal extends Modal<Events, ChallengeModalProperties, any> {
+export class ChallengeModal extends Modal<{}, ChallengeModalProperties, any> {
     constructor(props: ChallengeModalProperties) {
         super(props);
     }
@@ -2142,340 +2069,6 @@ export class ChallengeModalBody extends React.Component<
     };
 }
 
-export function challenge(
-    player_id?: number,
-    initial_state?: any,
-    computer?: boolean,
-    config?: ChallengeModalConfig,
-    created?: (c: CreatedChallengeInfo) => void,
-) {
-    // TODO: Support challenge by player, w/ initial state, or computer
-
-    if (player_id && typeof player_id !== "number") {
-        console.log("Invalid player id: ", player_id);
-        throw Error("Invalid player id");
-    }
-
-    let mode: ChallengeModes = "open";
-    if (player_id) {
-        mode = "player";
-    }
-    if (computer) {
-        mode = "computer";
-    }
-
-    return openModal(
-        <ChallengeModal
-            playerId={player_id}
-            initialState={initial_state}
-            config={config}
-            mode={mode}
-            created={created}
-        />,
-    );
-}
-export function createGameRecord(props: {
-    library_collection_id?: number;
-    players_list?: Array<{ name: string; rank: number }>;
-    tournament_record_id?: number;
-    tournament_record_round_id?: number;
-}) {
-    const mode: ChallengeModes = "demo";
-    return openModal(
-        <ChallengeModal
-            mode={mode}
-            game_record_mode={true}
-            libraryCollectionId={props.library_collection_id}
-            playersList={props.players_list}
-            tournamentRecordId={props.tournament_record_id}
-            tournamentRecordRoundId={props.tournament_record_round_id}
-        />,
-    );
-}
-export function createDemoBoard(
-    players_list?: Array<{ name: string; rank: number }>,
-    tournament_record_id?: number,
-    tournament_record_round_id?: number,
-) {
-    const mode: ChallengeModes = "demo";
-    return openModal(
-        <ChallengeModal
-            mode={mode}
-            playersList={players_list}
-            tournamentRecordId={tournament_record_id}
-            tournamentRecordRoundId={tournament_record_round_id}
-        />,
-    );
-}
-
-export function challengeComputer(settings?: ChallengeModalConfig) {
-    return challenge(undefined, null, true, settings);
-}
-export function challengeRematch(
-    goban: GobanRenderer,
-    opponent: GobanEnginePlayerEntry,
-    original_game_meta: GobanEngineConfig & { pause_on_weekends?: boolean },
-) {
-    /* TODO: Fix up challengeRematch time control stuff */
-    const conf = goban.engine;
-
-    const board_size = `${conf.width}x${conf.height}`;
-
-    console.log(original_game_meta);
-
-    //config.syncBoardSize();
-    //config.syncTimeControl();
-
-    const config: ChallengeModalConfig = {
-        conf: {
-            selected_board_size: isStandardBoardSize(board_size) ? board_size : "custom",
-            restrict_rank: false,
-        },
-        challenge: {
-            challenger_color:
-                conf.players.black.id === opponent.id ? ("white" as const) : ("black" as const),
-            game: {
-                handicap: conf.handicap,
-                time_control: conf.time_control,
-                rules: conf.rules,
-                ranked: !!conf.config.ranked,
-                width: conf.width,
-                height: conf.height,
-                komi_auto: "custom",
-                komi: conf.komi,
-                disable_analysis: conf.disable_analysis,
-                pause_on_weekends: original_game_meta.pause_on_weekends ?? false,
-                initial_state: null,
-                private: (conf as any)["private"], // this is either missing from the type def or invalid
-            },
-            invite_only: false, // maybe one day we will support challenge-links on the rematch dialog!
-        },
-        time_control: dup(conf.time_control),
-    };
-
-    challenge(opponent.id, null, false, config);
-}
-export function challenge_text_description(challenge: ChallengeDetails) {
-    const c = challenge;
-    const g = "game" in challenge ? challenge.game : challenge;
-    let challenge_description = g.ranked ? _("Ranked") : _("Unranked");
-
-    if (g.rengo) {
-        challenge_description +=
-            " " + (g.rengo_casual_mode ? _("casual rengo") : _("strict rengo"));
-        if (c.rengo_auto_start > 0) {
-            challenge_description +=
-                ", " +
-                interpolate(_("auto-starting at {{auto_start}} players"), {
-                    auto_start: c.rengo_auto_start,
-                });
-        }
-    }
-
-    challenge_description +=
-        ", " + g.width + "x" + g.height + ", " + interpolate(_("%s rules"), [rulesText(g.rules)]);
-
-    const time_control =
-        typeof g.time_control !== "object" && "time_control_parameters" in g
-            ? g.time_control_parameters
-            : g.time_control;
-
-    if (typeof time_control === "object") {
-        if (time_control.system !== "none") {
-            // This is gross and still needs refactoring, but preserves the old behavior
-            const time_control_system = time_control.system;
-            challenge_description +=
-                ", " +
-                interpolate(_("%s clock: %s"), [
-                    timeControlSystemText(time_control_system).toLocaleLowerCase(),
-                    shortShortTimeControl(time_control),
-                ]);
-        } else {
-            challenge_description += ", " + _("no time limits");
-        }
-
-        if (time_control.pause_on_weekends) {
-            challenge_description += ", " + _("pause on weekends");
-        }
-    }
-
-    challenge_description +=
-        ", " +
-        interpolate(_("%s handicap"), [g.handicap < 0 ? _("auto") : g.handicap]) +
-        (g?.komi_auto !== "custom" || g.komi == null || typeof g.komi === "object"
-            ? ""
-            : ", " + interpolate(_("{{komi}} komi"), { komi: g.komi })) +
-        (g.disable_analysis ? ", " + _("analysis disabled") : "");
-
-    // (In some places we use +/-1000 to represent "no rank limit)"
-
-    if (
-        c.min_ranking !== undefined &&
-        c.min_ranking > 0 &&
-        c.max_ranking !== undefined &&
-        c.max_ranking < 1000
-    ) {
-        challenge_description +=
-            ", " +
-            interpolate(_("ranks {{min}} to {{max}}"), {
-                min: rankSelectorIndexToText(c.min_ranking),
-                max: rankSelectorIndexToText(c.max_ranking),
-            });
-    }
-
-    if (c.challenger_color !== "automatic") {
-        let your_color = "";
-
-        const challenger_id = (c as any)?.challenger?.id || (c as any)?.user?.id;
-        if (challenger_id && challenger_id !== data.get("user")?.id) {
-            if (c.challenger_color === "black") {
-                your_color = _("white");
-            } else if (c.challenger_color === "white") {
-                your_color = _("black");
-            } else {
-                your_color = _(c.challenger_color);
-            }
-        } else {
-            your_color = _(c.challenger_color);
-        }
-
-        challenge_description +=
-            ", " + interpolate(pgettext("color", "you play as %s"), [your_color]);
-    }
-
-    return challenge_description;
-}
-
-function isStandardBoardSize(board_size: string): boolean {
+export function isStandardBoardSize(board_size: string): boolean {
     return board_size in standard_board_sizes;
-}
-
-export interface ChallengeModalConfig {
-    challenge: {
-        min_ranking?: number;
-        max_ranking?: number;
-        challenger_color: rest_api.ColorSelectionOptions;
-        invite_only: boolean;
-
-        game: {
-            name?: string;
-            rules: RuleSet;
-            ranked: boolean;
-            handicap: number;
-            komi_auto: string;
-            disable_analysis: boolean;
-            initial_state: GobanEngineInitialState | null;
-            private: boolean;
-            time_control?: JGOFTimeControl;
-            width?: number;
-            height?: number;
-            komi?: number;
-            pause_on_weekends?: boolean;
-        };
-    };
-    conf: {
-        restrict_rank: boolean;
-        selected_board_size?: string;
-    };
-    time_control: TimeControlConfig;
-}
-
-interface TimeControlConfig {
-    system: JGOFTimeControlSystem;
-    speed: JGOFTimeControlSpeed;
-    initial_time?: number;
-    main_time?: number;
-    time_increment?: number;
-    max_time?: number;
-    period_time?: number;
-    periods?: number;
-    pause_on_weekends: boolean;
-}
-
-/* This function provides translations for rejection reasons coming gtp2ogs bot interface scripts. */
-export function rejectionDetailsToMessage(details: RejectionDetails): string | undefined {
-    switch (details.rejection_code) {
-        case "blacklisted":
-            return pgettext(
-                "The user has been blocked by the operator of the bot from playing against the bot.",
-                "The operator of this bot will not let you play against it.",
-            );
-
-        case "board_size_not_square":
-            return _("This bot only plays on square boards");
-
-        case "board_size_not_allowed":
-            return _("The selected board size is not supported by this bot");
-
-        case "handicap_not_allowed":
-            return _("Handicap games are not allowed against this bot");
-
-        case "unranked_not_allowed":
-            return _("Unranked games are not allowed against this bot");
-
-        case "ranked_not_allowed":
-            return _("Ranked games are not allowed against this bot");
-
-        case "blitz_not_allowed":
-            return _("Blitz games are not allowed against this bot");
-
-        case "too_many_blitz_games":
-            return _(
-                "Too many blitz games are being played by this bot right now, please try again later.",
-            );
-
-        case "live_not_allowed":
-            return _("Live games are not allowed against this bot");
-
-        case "too_many_live_games":
-            return _(
-                "Too many live games are being played by this bot right now, please try again later.",
-            );
-
-        case "correspondence_not_allowed":
-            return _("Correspondence games are not allowed against this bot");
-
-        case "too_many_correspondence_games":
-            return _(
-                "Too many correspondence games are being played by this bot right now, please try again later.",
-            );
-
-        case "time_control_system_not_allowed":
-            return _("The provided time control system is not supported by this bot");
-
-        case "time_increment_out_of_range":
-            return _("The time increment is out of the acceptable range allowed by this bot");
-
-        case "period_time_out_of_range":
-            return _("The period time is out of the acceptable range allowed by this bot");
-
-        case "periods_out_of_range":
-            return _("The number of periods is out of the acceptable range allowed by this bot");
-
-        case "main_time_out_of_range":
-            return _("The main time is out of the acceptable range allowed by this bot");
-
-        case "max_time_out_of_range":
-            return _("The max time is out of the acceptable range allowed by this bot");
-
-        case "per_move_time_out_of_range":
-            return _("The per move time is out the acceptable range allowed by this bot");
-
-        case "player_rank_out_of_range":
-            return _("Your rank is too high or low to play against this bot");
-
-        case "not_accepting_new_challenges":
-            return _("This bot is not accepting new games at this time");
-
-        case "too_many_games_for_player":
-            return _(
-                "You are already playing against this bot, please end your other game before starting a new one",
-            );
-
-        case "komi_out_of_range":
-            return _("Komi is out of the acceptable range allowed by this bot");
-
-        default:
-            return undefined;
-    }
 }
